requirements.txt
# Core data science packages
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0

# Visualization packages
plotly>=5.10.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Machine learning packages
scikit-learn>=1.1.0
joblib>=1.2.0

# Dashboard packages
streamlit>=1.15.0
dash>=2.6.0
dash-bootstrap-components>=1.2.0

# Data processing packages
openpyxl>=3.0.0
xlsxwriter>=3.0.0

# Utility packages
pathlib2>=2.3.0
python-dateutil>=2.8.0
pytz>=2022.1

# Configuration packages
pyyaml>=6.0
configparser>=5.3.0

# Testing packages
pytest>=7.0.0
pytest-cov>=4.0.0

# Image export (optional)
kaleido>=0.2.1

# Development packages
jupyterlab>=3.4.0
notebook>=6.4.0
